requirements.txt
opencv-python>=4.13.0
numpy>=2.0.0
pillow>=11.0.0
